refactor(fetch-all-networks): split into network-fetcher/{types,fetch}
Pure-TDD structural split of the 610-line `lib/fetch-all-networks.ts` into
two focused modules. The original path stays as a barrel re-export so all
7 importers continue to work unchanged.

- `lib/network-fetcher/types.ts` (85 lines) — `NetworkData`,
  `SnapshotPageResult`, plus `SnapshotWindows`/`TimeRange` re-exports from
  `@/lib/volume`. Type-only consumers no longer transitively pull
  `graphql-request` and `@sentry/nextjs`.
- `lib/network-fetcher/fetch.ts` (548 lines) — `fetchNetworkData`,
  `fetchAllNetworks`, `blankNetworkData`, `isNetworkDataFullyHealthy`,
  `REQUEST_TIMEOUT_MS`, mutable module state (`warnedCapKeys`,
  `partialPageLastCapturedAt`), plus the private pagination helpers.
- `lib/fetch-all-networks.ts` (9 lines) — barrel-only.

TDD safety net: added `lib/__tests__/fetch-all-networks.test.ts` (mirrors
the `EXPECTED_EXPORT_NAMES` pattern from `queries.test.ts`) BEFORE the move
and ran it green against the un-split file. The integration test at
`hooks/__tests__/use-all-networks-data.test.ts` (26 tests) is the
characterization gate and stays green through the split.

Test: 1472 passed (was 1466; +6 surface-contract assertions).
Typecheck: clean.
Importers: 7, unchanged (verified via grep).
